requirements.txt
# Python dependencies for Garmin integration
garminconnect==0.2.19
cloudscraper==1.2.71
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0